refact: avoid repeating code in `updateRepos`
This change should give the exact same result, except moving the `else`-block within the try/catch (and I don't see why that is bad).

The refactor makes it easier to understand the difference in logic with and without a present `repoConfig` by only having the actual alternate action within the if/else and keeping the bit that always happen outside the if/else.