chore: autoformatted after changed eslint config in #664
Mostly, the changes in #664 relaxes the eslint rules, but that PR also removes the space before the parentheses in function definitions.
The formatting was done with the vscode js/ts formatter `vscode.typescript-language-features`, and since this also does the same change it is now consistent across the whole codebase.

Other than removing the space-before-function-paren, some files with inconsistent indentation and/or trailing spaces has been updated.