refact: extracted result handling to a separate class
Also moved the `handleError` function from settings.js to index.js.

The rationale for this change is that result handling sometimes happens outside the context of an instance of the settings class (as evinced by the `handleError` function).
I believe this change improves the separation of concerns and thus also makes the code more readable.